Revert "majorly refactored server.js, split routes into diff files"
