--- cypress.config.js
@@ -1,7 +1,0 @@
-const { defineConfig } = require('cypress')
-
-module.exports = defineConfig({
-  e2e: {
-    baseUrl: 'http://localhost:8080',
-  },
-})
