+++ package.json
@@ -4,10 +4,11 @@
     "description": "",
     "type": "commonjs",
     "scripts": {
+        "devStart": "nodemon src/server.js",
         "start": "node src/server.js",
         "dev": "nodemon src/server.js",
         "test": "jest",
-        "cypress": "cypress open"
+        "testcov": "jest --coverage"
     },
     "repository": {
         "type": "git",
@@ -29,7 +30,6 @@
         "ejs": "^3.1.8",
         "express": "^4.18.2",
         "express-flash": "^0.0.2",
-        "express-rate-limit": "^6.7.0",
         "express-session": "^1.17.3",
         "express-validator": "^6.15.0",
         "h5p-nodejs-library": "^6.2.0",
@@ -41,7 +41,6 @@
         "supertest": "^6.3.3"
     },
     "devDependencies": {
-        "cypress": "^12.7.0",
         "dotenv": "^16.0.3",
         "eslint": "^8.35.0",
         "eslint-config-standard": "^17.0.0",
