+++ src/server.js
@@ -1,64 +1,47 @@
-
-/**
- * modular dependancies
-*/
-const express = require("express")
-const server = express()
-//const render  = require("ejs") //todo, not sure if this is needed in this file
-const flash = require("express-flash")
-const session = require("express-session")
-const passport = require("passport")
-//const { check, validationResult } = require("express-validator") //todo, not sure if this is needed in this file
-const methodOverride = require("method-override")
-//const sendEmail = require("../utils/email/sendEmail") //todo, not sure if this is needed in this file
-const rateLimit = require('express-rate-limit')
-
-const { Prisma } = require("@prisma/client") //TODO: is this needed/being used here, also see database.js
-const { application } = require("express") //TODO: is this being used?
-
-
-
-const ip = require('../utils/getPublicIp')
-const auth = require('./authenticate')
-const regestrationRoute = require('./registration')
-const usersRoute = require('./users')
-const accountRoute = require('./account')
-const passwordRoute = require('./password')
-const initializePassport = require("./config/passport")
-const db = require("./config/database")
-server.set('view engine', 'ejs');
-
-
-
-/***
- * Config
- */
-if (process.env.NODE_ENV !== "production") {//checks if we are in prod envoirment
+if (process.env.NODE_ENV !== "production") {
     require("dotenv").config()
 }
+
+const express = require("express")
+const db = require("./config/database")
+const bcrypt = require("bcrypt")
+const { Prisma } = require("@prisma/client")
+const initializePassport = require("./config/passport")
+const flash = require("express-flash")
+const session = require("express-session")
+const { application } = require("express")
+const passport = require("passport")
+const methodOverride = require("method-override")
+const jwt = require("jsonwebtoken")
+const { check, validationResult } = require("express-validator")
+const server = express()
+const sendEmail = require("../utils/email/sendEmail")
+const { render } = require("ejs")
+
+server.set('view engine', 'ejs');
+
+const salt = bcrypt.genSaltSync(10);
+const JWT_SECRET = process.env.JWT_SECRET
 initializePassport(
     passport,
     async email => await db.User.findFirst({ where: { email } }),
     async id => await db.User.findFirst({ where: { id } })
 )
+const PAGE_SIZE = 8;
+server.use('/public', express.static('public'));
+// below line of code is to get the form data in req.body
+server.use(express.urlencoded({ extended: false }))
+server.use(flash())
 server.use(session({
     secret: process.env.SESSION_SECRET,
+
     resave: false, // we want to resave the session variable if nothing is changed
     saveUninitialized: false
 }))
-const limiter = rateLimit({
-    windowMs: 1*60*1000, //1 minute
-    max: 10
-})
-//server.use(limiter) //todo, enable, currently disabled to test something
-server.set('views', './views')
-server.set('view engine', 'ejs')
-server.use('/public', express.static('public')) //TODO: may be expressed/condensed as app.use(express.static('public')) see here: https://expressjs.com/en/starter/static-files.html
-server.use(express.urlencoded({ extended: false }))
-server.use(flash())
 server.use(passport.initialize())
 server.use(passport.session())
 server.use(methodOverride("_method"))
+
 async function main() {
     const PORT = 8080
 
@@ -218,18 +201,8 @@ server.get('/reset-password/:id/:token', async(req, res) => {
         }
     }
 
-/**
- * server code, primarily uses Expresses routes, and creates 'mini-apps' for the main functionalities
- * of our application
- */
-server.get('/', async(req, res) => {
-    res.render("index.ejs")
 })
 
-server.use('/registration', regestrationRoute)
-server.use('/users', usersRoute)
-server.use('/account', accountRoute)
-server.use('/password', passwordRoute)
 server.post('/reset-password/:id/:token',
     check('confirmPassword').custom((value, { req }) => {
         if (value !== req.body.password) {
@@ -294,16 +267,142 @@ server.get('/users', checkAuthenticated, async(req, res) => {
     const count = await db.user.count()
     const totalPages = Math.ceil(count / PAGE_SIZE)
 
-//ip() //posts our public IP to the console
+    try {
+        const user = await db.User.findFirst({
+            where: { id: req.user.id },
+            include: { role: true },
+        })
+
+        //console.log(user)
+
+        if (!user || user.role[0].role !== "Admin") {
+            // req.flash("error", "You do not have access to view users.")
+            res.render("unauthorized.ejs")
+            return
+        }
+
+
+        const users = await db.user.findMany({
+            skip: skip,
+            take: limit,
+            select: {
+                id: true,
+                email: true,
+                firstName: true,
+                lastName: true,
+            },
+        });
+
+        console.log("users length is " + count)
+        res.render("users.ejs", { users, page, totalPages });
+    } catch (error) {
+        console.error(error);
+        res.status(500).send('Internal server error');
+    }
+});
+
+server.post('/users/:id/delete', async(req, res) => {
+    const userId = parseInt(req.params.id);
+    console.log(userId)
+    try {
+        // Find the user by ID
+        const user = await db.user.findUnique({
+            where: {
+                id: userId,
+            },
+        });
+        console.log(user);
+        if (!user) {
+            return res.status(404).send('User not found');
+        }
+        // Delete the user
+        await db.user.delete({
+            where: {
+                id: userId,
+            },
+        });
+
+        // Redirect to the list of users
+        res.redirect('/users');
+    } catch (error) {
+        console.error(error);
+        res.status(500).send('Internal server error');
+    }
+});
+
+server.get('/users/:id/edit', async(req, res) => {
+    try {
+        const userId = parseInt(req.params.id);
+        const user = await db.user.findUnique({
+            where: {
+                id: userId
+            },
+            include: {
+                role: true
+            }
+        });
+        console.log(user)
+        if (!user) {
+            return res.status(404).send('User not found');
+        }
+
+        res.render('edit-user', {
+            user: user
+        });
+    } catch (error) {
+        console.error(error);
+        res.status(500).send('Internal server error');
+    }
+});
+
+server.post('/users/:id/edit', async(req, res) => {
+    const { firstName, lastName, email, role } = req.body;
+    const userId = parseInt(req.params.id);
+
+    try {
+        const updatedUser = await db.user.update({
+            where: {
+                id: userId
+            },
+            data: {
+                firstName: firstName,
+                lastName: lastName,
+                email: email
+            }
+        });
+        const updatedUserRole = await db.userRole.update({
+            where: {
+                id: userId
+            },
+            data: {
+                role: role,
+            },
+        });
+
+        req.flash('success', 'User updated successfully!');
+        res.redirect('/users');
+    } catch (err) {
+        console.error(err);
+        req.flash('error', 'Failed to update user.');
+        res.redirect('/users');
+    }
+});
 
 
 
+function checkAuthenticated(req, res, next) {
+    if (req.isAuthenticated()) {
+        return next()
+    }
+    res.redirect("/login")
+}
 
+function checkNotAuthenticated(req, res, next) {
+    if (req.isAuthenticated()) {
+        return res.redirect("/create")
+    }
+    next()
+}
+main();
 
-
-
-
-
-
-
-
+module.exports = server
