+++ views/create.ejs
@@ -32,13 +32,12 @@
     </head>
 
     <body>
-        <%- include('../views/partial/_sitemap.ejs'); %>
         <div class="main">
             <div class="form-container">
                 <h2 class="text-center">Welcome,
                     <%= name %>
                 </h2>
-                <form action="/account/logout?_method=DELETE" method="POST">
+                <form action="/logout?_method=DELETE" method="POST">
                     <div class="text-center">
                         <button type="submit" class="btn btn-danger btn-rounded btn-lg">Logout</button>
                     </div>
