+++ views/forgotPassword.ejs
@@ -24,6 +24,7 @@
             <% if(messages.error) { %>
             <div class="alert alert-danger" role="alert">
               <strong><%= messages.error %></strong>
+
             </div>
             <% } %>
             <form action="/forgot-password" method="POST">
