+++ views/index.ejs
@@ -14,7 +14,7 @@
   </head>
   <body>
     <%- include('../views/partial/_navHeader.ejs'); %>
-    <%- include('../views/partial/_sitemap.ejs')
+
     <!--Hero Section-->
     <div class="main">
       <div class="main__container">
