+++ views/login.ejs
@@ -49,7 +49,6 @@
                         />
                       </div>
 
-
                       <div class="form-outline mb-4">
                         <input
                           type="password"
