+++ views/partial/_header.ejs
@@ -1,0 +1,11 @@
+<nav class="header-nav">
+    <header>
+        <div class="header-container">
+            <ul>
+                <a href="#">Home</a>
+                <a href="#">About</a>
+                <a href="#">Contact</a>
+            </ul>
+        </div>
+    </header>
+</nav>
