+++ views/partial/_navHeader.ejs
@@ -18,7 +18,7 @@
         <a href="/" class="navbar__links"> Demo </a>
       </li>
       <li class="navbar__btn">
-        <a href="/account/login" class="button"> Login </a>
+        <a href="/login" class="button"> Login </a>
       </li>
     </ul>
   </div>
