+++ views/partial/_sitemap.ejs
@@ -1,10 +1,9 @@
 <div class="site-links">
     <h3>bunch of links to help navigate through the site</h3>
-    <a href="../account/homepage">User homepage</a>
-    <a href="../password/forgot">Forgot password</a>
-    <a href="../">Index</a>
-    <a href="../account/login">Login</a>
-    <a href="../registration">Registration</a>
-    <a href="../users">Users</a>
-
+    <a href="../create">create</a>
+    <a href="../forgot-password">forgot password</a>
+    <a href="../">index</a>
+    <a href="../login">login</a>
+    <a href="../registration">registration</a>
+    <a href="../reset-password">resetPassword</a>
 </div>
