+++ views/registration.ejs
@@ -18,6 +18,7 @@
   <body>
     <!--Navigation EJS Partial call do not remove-->
     <%- include('../views/partial/_navHeader.ejs'); %>
+
     <section class="h-100 gradient-form">
       <div class="container py-1 h-100">
         <div class="row d-flex justify-content-center align-items-center h-100">
@@ -29,7 +30,6 @@
                     <div class="text-center">
                       <h4 class="mt-5 mb-1 pb-2">A U G M E N T</h4>
                     </div>
-
 
                     <form
                       action="/registration"
