--- views/tempReg.ejs
@@ -1,51 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-</head>
-<body>
-    <%- include('../views/partial/_navHeader.ejs'); %>
-
-    <form action="/registration" method="POST" class="input-form">
-
-        
-
-                <h5 class="text-center" style="padding: 20px;"><b>Register</b></h5>
-                <div class="form-group">
-                    <input type="text" name="firstName" class="form-control" id="firstName" placeholder="First name">
-                </div>
-                <div class="form-group">
-                    <input type="text" name="lastName" class="form-control" id="lastName" placeholder="Last name">
-                </div>
-                <div class="form-group">
-                    <input type="email" name="email" class="form-control" id="email" placeholder="Email">
-                </div>
-                <div class="form-group">
-                    <input type="password" name="password" class="form-control" id="password" placeholder="Password">
-                </div>
-                <div class="form-group">
-                    <input type="password" name="confirmPassword" class="form-control" id="confirmPassword" placeholder="Confirm Password">
-                </div>
-                <div class="form-group-form-check">
-                    <input type="checkbox" class="form-check-input" style="float:left;" id="remember-me">
-                    <label class="text-center" style="float:left; padding-left: 8px;" for="remember-me">Remember
-            me</label>
-                </div>
-                <div class="form-group-lgn-btn">
-                    <button type="submit" class="btn btn-outline-success" onclick="">Register</button>
-                </div>
-                <div>
-                    <label class="text-center" style="text-align:center; padding-left: 8px; margin-top: 20px;">Click
-            here to</label>
-                    <a href="/login" style="text-align:center; margin-top: 20px; padding-left: 5px;">Login</button></a>
-                </div>
-    </form>
-
-</div>
-
-</form>
-</body>
-</html>
